+++ lib/csv.dart
@@ -1,15 +1,16 @@
 library csv;
 
-import 'dart:convert';
 import 'dart:async';
+import 'dart:convert';
 
-import 'src/csv_parser.dart';
-import 'src/complex_converter.dart';
 import 'csv_settings_autodetection.dart';
+import 'src/complex_converter.dart';
+import 'src/csv_parser.dart';
 
 part 'csv_to_list_converter.dart';
-
 part 'list_to_csv_converter.dart';
+
+enum EmptyValue { NULL }
 
 /// The RFC conform default value for field delimiter.
 const String defaultFieldDelimiter = ',';
@@ -35,18 +36,22 @@ class CsvCodec {
       String eol = defaultEol,
       bool shouldParseNumbers = true,
       bool allowInvalid = true,
-      bool delimitAllFields = defaultDelimitAllFields})
+      bool delimitAllFields = defaultDelimitAllFields,
+      convertNullTo,
+      convertEmptyTo})
       : decoder = CsvToListConverter(
             fieldDelimiter: fieldDelimiter,
             textDelimiter: textDelimiter,
             textEndDelimiter: textEndDelimiter,
             eol: eol,
             shouldParseNumbers: shouldParseNumbers,
-            allowInvalid: allowInvalid),
+            allowInvalid: allowInvalid,
+            convertEmptyTo: convertEmptyTo),
         encoder = ListToCsvConverter(
             fieldDelimiter: fieldDelimiter,
             textDelimiter: textDelimiter,
             textEndDelimiter: textEndDelimiter,
             eol: eol,
-            delimitAllFields: delimitAllFields);
+            delimitAllFields: delimitAllFields,
+            convertNullTo: convertNullTo);
 }